code_main/requirements.txt
pytesseract
opencv-python-headless
boto3
numpy
pillow
spacy
transformers
torch
sentencepiece
symspellpy
amazon-textract-response-parser
amazon-textract-caller
fuzzywuzzy
gradio
pandas
requests
matplotlib
scikit-learn
tabulate
google-generativeai
textractor
rapidfuzz
